Use long for command timeout and document units.
Using a long is standard practice and also allows for using the TimeUnit classes
to provide declarative values (e.g., `TimeUnit.SECONDS.toMillis(30)`).

A special circle of hell is reserved for those who have method arguments for timeouts
but do not document the units. All classes accepting a timeout argument now explicitly
state they accept milliseconds.

+++ ddmlib/src/main/java/com/android/ddmlib/testrunner/IRemoteAndroidTestRunner.java
@@ -174,7 +174,7 @@ public interface IRemoteAndroidTestRunner {
 
     /**
      * Sets the maximum time allowed between output of the shell command running the tests on
-     * the devices.
+     * the devices (in milliseconds).
      * <p/>
      * This allows setting a timeout in case the tests can become stuck and never finish. This is
      * different from the normal timeout on the connection.
@@ -183,7 +183,7 @@ public interface IRemoteAndroidTestRunner {
      *
      * @see IDevice#executeShellCommand(String, com.android.ddmlib.IShellOutputReceiver, int)
      */
-    public void setMaxtimeToOutputResponse(int maxTimeToOutputResponse);
+    public void setMaxtimeToOutputResponse(long maxTimeToOutputResponse);
 
     /**
      * Set a custom run name to be reported to the {@link ITestRunListener} on {@link #run}
